feat: add select and enhance update
